Strings for Table Added
Added Strings for the WMS table
Cleaned up main activity
