+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,54 @@
 <resources>
     <string name="app_name">Wild Magic Surge</string>
 
-    <string name="dummy_button">Dummy Button</string>
-    <string name="dummy_content">DUMMY\nCONTENT</string>
+    <string name="Surge01_02">Roll on this table at the start of each of your turns for the next minute, ignoring this result on subsequent rolls.</string>
+    <string name="Surge03_04">For the next minute, you can see any Invisible creature if you have line of sight to it.</string>
+    <string name="Surge05_06">A modron chosen and controlled by the DM appears in an unoccupied space within 5 feet of you, then disappears 1 minute later.</string>
+    <string name="Surge07_08">You cast Fireball as a 3rd-level spell centered on yourself.</string>
+    <string name="Surge09_10">You cast Magic Missile as a 5th-level spell.</string>
+    <string name="Surge11_12">Roll a d10. Your height changes by a number of inches equal to the roll. If the roll is odd, you shrink. If the roll is even, you grow.</string>
+    <string name="Surge13_14">You cast Confusion centered on yourself.</string>
+    <string name="Surge15_16">For the next minute, you regain 5 Hit Points at the start of each of your turns.</string>
+    <string name="Surge17_18">You grow a long beard made of feathers that remains until you sneeze, at which point the feathers explode out from your face.</string>
+    <string name="Surge19_20">You cast Grease centered on yourself.</string>
+    <string name="Surge21_22">Creatures have disadvantage on Saving Throws against the next spell you cast in the next minute that involves a saving throw.</string>
+    <string name="Surge23_24">Your skin turns a vibrant shade of blue. A Remove Curse spell can end this effect.</string>
+    <string name="Surge25_26">An eye appears on your forehead for the next minute. During that time, you have advantage on Wisdom (Perception) checks that rely on sight.</string>
+    <string name="Surge27_28">For the next minute, all your Spells with a Casting Time of 1 action have a Casting Time of 1 Bonus Action.</string>
+    <string name="Surge29_30">You Teleport up to 60 feet to an unoccupied space of your choice that you can see.</string>
+    <string name="Surge31_32">You are transported to the Astral Plane until the end of your next turn, after which time you return to the space you previously occupied or the nearest unoccupied space if that space is occupied.</string>
+    <string name="Surge33_34">Maximize the damage of the next damaging spell you cast within the next minute.</string>
+    <string name="Surge35_36">Roll a d10. Your age changes by a number of years equal to the roll. If the roll is odd, you get younger (minimum 1 year old). If the roll is even, you get older.</string>
+    <string name="Surge37_38">1d6 flumphs controlled by the DM appear in unoccupied spaces within 60 feet of you and are Frightened of you. They Vanish after 1 minute.</string>
+    <string name="Surge39_40">You regain 2d10 Hit Points.</string>
+    <string name="Surge41_42">You turn into a potted plant until the start of your next turn. While a plant, you are Incapacitated and have vulnerability to all damage. If you drop to 0 Hit Points, your pot breaks, and your form reverts.</string>
+    <string name="Surge43_44">For the next minute, you can teleport⁠ up to 20 feet as a Bonus Action on each of your turns.</string>
+    <string name="Surge45_46">You cast Levitate on yourself.</string>
+    <string name="Surge47_48">A Unicorn controlled by the DM appears in a space within 5 feet of you, then disappears 1 minute later.</string>
+    <string name="Surge49_50">You can’t speak for the next minute. Whenever you try, pink bubbles float out of your mouth.</string>
+    <string name="Surge51_52">A spectral shield⁠ hovers near you for the next minute, granting you a +2 bonus to AC and immunity to Magic Missile.</string>
+    <string name="Surge53_54">You are immune to being intoxicated by alcohol for the next 5d6 days.</string>
+    <string name="Surge55_56">Your hair falls out but grows back within 24 hours.</string>
+    <string name="Surge57_58">For the next minute, any flammable object you touch that isn’t being worn or carried by another creature bursts into flame.</string>
+    <string name="Surge59_60">You regain your lowest-level expended spell slot.</string>
+    <string name="Surge61_62">For the next minute, you must shout when you speak.</string>
+    <string name="Surge63_64">You cast Fog Cloud centered on yourself.</string>
+    <string name="Surge65_66">Up to three creatures you choose within 30 feet of you take 4d10 lightning damage.</string>
+    <string name="Surge67_68">You are Frightened by the nearest creature until the end of your next turn.</string>
+    <string name="Surge69_70">Each creature within 30 feet of you becomes Invisible for the next minute. The Invisibility ends on a creature when it attacks or casts a spell.</string>
+    <string name="Surge71_72">You gain Resistance to all damage for the next minute.</string>
+    <string name="Surge73_74">A random creature within 60 feet of you becomes Poisoned for 1d4 hours.</string>
+    <string name="Surge75_76">You glow with bright light in a 30-foot radius for the next minute. Any creature that ends its turn within 5 feet of you is Blinded until the end of its next turn.</string>
+    <string name="Surge77_78">You cast Polymorph on yourself. If you fail the saving throw, you turn into a sheep for the spell’s Duration.</string>
+    <string name="Surge79_80">Illusory butterflies and flower petals flutter in the air within 10 feet of you for the next minute.</string>
+    <string name="Surge81_82">You can take one additional action immediately.</string>
+    <string name="Surge83_84">Each creature within 30 feet of you takes 1d10 necrotic damage. You regain Hit Points equal to the sum of the necrotic damage dealt.</string>
+    <string name="Surge85_86">You cast Mirror Image.</string>
+    <string name="Surge87_88">You cast fly on a random creature within 60 feet of you.</string>
+    <string name="Surge89_90">You become Invisible for the next minute. During that time, other creatures can’t hear you. The Invisibility ends if you Attack or Cast a Spell.</string>
+    <string name="Surge91_92">If you die within the next minute, you immediately come back to life as if by the Reincarnate spell.</string>
+    <string name="Surge93_94">Your size increases by one size category for the next minute.</string>
+    <string name="Surge95_96">You and all creatures within 30 feet of you gain vulnerability to piercing damage for the next minute.</string>
+    <string name="Surge97_98">You are surrounded by faint, ethereal music for the next minute.</string>
+    <string name="Surge99_00">You regain all expended Sorcery Points.</string>
 </resources>
